requirements.txt
azure-functions
azure-storage-blob==12.16.0
azure-identity
python-dotenv==1.0.0
